refactor/movie-details: improve data handling
