+++ domain/src/main/java/com/london/domain/entity/moviedatails/CollectionDetails.kt
@@ -1,0 +1,6 @@
+package com.london.domain.entity.moviedatails
+
+data class CollectionDetails(
+    val id: Int,
+    val name: String
+)
