+++ domain/src/main/java/com/london/domain/entity/moviedatails/Genre.kt
@@ -1,0 +1,6 @@
+package com.london.domain.entity.moviedatails
+
+data class Genre(
+    val id: Int,
+    val name: String
+)
