+++ domain/src/main/java/com/london/domain/entity/moviedatails/ProductionCompany.kt
@@ -1,0 +1,8 @@
+package com.london.domain.entity.moviedatails
+
+data class ProductionCompany(
+    val id: Int,
+    val logoPath: String,
+    val name: String,
+    val originCountry: String
+)
