+++ domain/src/main/java/com/london/domain/entity/moviedatails/ProductionCountry.kt
@@ -1,0 +1,6 @@
+package com.london.domain.entity.moviedatails
+
+data class ProductionCountry(
+    val iso31661: String,
+    val name: String
+)
