+++ domain/src/main/java/com/london/domain/entity/moviedatails/SpokenLanguage.kt
@@ -1,0 +1,7 @@
+package com.london.domain.entity.moviedatails
+
+data class SpokenLanguage(
+    val englishName: String,
+    val iso6391: String,
+    val name: String
+)
